Make mh_export.py's download files even when the destination directory exists
We assumed before that the elements should be in their own folder, but this is
not the case for the different quality copies of a adaptive streaming video.

Therefore, now mh_export.py does not complain and refuse to copy a file when
the destination folder already exist. It does complain, however, when the
*file* already exists, because that should never happen.

Additionally, an internal flag was added to leave the smil files undeleted
after downloading all the video files.