(mh_clean_unarchived.py) Optimize first request by providing a page size 1
This first request is only used to get the total number of workflows, but no
search results are used here. By providing page size 0, Opencast assumes that
it should delete the default page size (probably 10). Even though this has very
low impact on fast networks, it still unnecessary to request ten elements when
we are not using any of them.