(SelectSeries.py) Modified the separator in the output file and remove argument
A different separator between the series ID and titles is now being used,
because we have many series with hyphens ('-') in the titles, but none with
colons (':'). Using a colon makes it easier to separate the fields in a script
or a program such as Excel.

The argument "all", that accepted a file name where all the read series could
be written to, has been removed. A user can get the same effect by selecting
all the series by clicking the corresponding button, then generating the file
with that selection.